backend/requirements.txt
mysql-connector-python
fastapi==0.89.1
uvicorn==0.20.0
pydantic
python-decouple == 3.3
fastapi-sessions
python-multipart

